Dep-eject progress.html.render
