+++ lib/depject/app/html/progress-notifier.js
@@ -7,7 +7,6 @@ exports.gives = nest('app.html.progressNotifier')
 
 exports.needs = nest({
   'sbot.pull.stream': 'first',
-  'progress.html.render': 'first',
   'progress.obs': {
     indexes: 'first',
     replicate: 'first',
