+++ lib/depject/index.js
@@ -167,8 +167,7 @@ module.exports = {
     },
     progress: {
       html: {
-        peer: require('./progress/html/peer.js'),
-        render: require('./progress/html/render.js')
+        peer: require('./progress/html/peer.js')
       },
       obs: require('./progress/obs.js')
     },
