+++ lib/depject/progress/html/peer.js
@@ -1,10 +1,10 @@
 const { computed, when } = require('mutant')
 const nest = require('depnest')
+const renderProgress = require('../../../progress/html/render')
 
 exports.gives = nest('progress.html.peer')
 
 exports.needs = nest({
-  'progress.html.render': 'first',
   'progress.obs.peer': 'first',
   'progress.obs.replicate': 'first'
 })
@@ -17,6 +17,6 @@ exports.create = function (api) {
       return (feeds - pending) / feeds
     })
 
-    return api.progress.html.render(value, when(progress, '-pending'))
+    return renderProgress(value, when(progress, '-pending'))
   })
 }
